Split crate into sev and snp modules
Due to the fact that certain commands are SEV specific, along with the
incoming SNP-specific commands, it's best to split this crate up into
two main modules, "sev" and "snp". However, some commands are able to be
run on both generations, so we can include them outside of the modules.
