+++ src/snp_cmds/mod.rs
@@ -1,0 +1,7 @@
+// SPDX-License-Identifier: Apache-2.0
+
+use anyhow::Result;
+
+pub fn cmd() -> Result<()> {
+    todo!();
+}
